refactor: Reduce redundancy by creating abstract gen instruction class for training and applying gen models
