+++ immuneML/workflows/instructions/train_gen_model/TrainGenModelInstruction.py
@@ -1,37 +1,16 @@
-import copy
-from abc import ABC
-from dataclasses import dataclass, field
 from pathlib import Path
-from typing import List, Dict
 
-from immuneML.IO.dataset_export.AIRRExporter import AIRRExporter
 from immuneML.data_model.dataset.Dataset import Dataset
-from immuneML.environment.SequenceType import SequenceType
 from immuneML.ml_methods.generative_models.GenerativeModel import GenerativeModel
-from immuneML.reports.ReportResult import ReportResult
-from immuneML.reports.data_reports.DataReport import DataReport
-from immuneML.reports.gen_model_reports.GenModelReport import GenModelReport
 from immuneML.util.Logger import print_log
-from immuneML.util.PathBuilder import PathBuilder
-from immuneML.workflows.instructions.Instruction import Instruction
-
-
-@dataclass
-class GenModelState(ABC):
-    result_path: Path
-    name: str
-    gen_examples_count: int
-    sequence_examples: list = None
-    model_path: Path = None
-    report_results: Dict[str, List[ReportResult]] = field(
-        default_factory=lambda: {'data_reports': [], 'ml_reports': []})
+from immuneML.workflows.instructions.GenModelInstruction import GenModelState, GenModelInstruction
 
 
 class TrainGenModelState(GenModelState):
     pass
 
 
-class TrainGenModelInstruction(Instruction):
+class TrainGenModelInstruction(GenModelInstruction):
     """
     TrainGenModel instruction implements training generative AIRR models on receptor level. Models that can be trained
     for sequence generation are listed under Generative Models section.
@@ -73,14 +52,11 @@ class TrainGenModelInstruction(Instruction):
 
     def __init__(self, dataset: Dataset = None, method: GenerativeModel = None, number_of_processes: int = 1,
                  gen_examples_count: int = 100, result_path: Path = None, name: str = None, reports: list = None):
+        super().__init__(TrainGenModelState(result_path, name, gen_examples_count), method, reports)
         self.dataset = dataset
-        self.generated_dataset = None
         self.number_of_processes = number_of_processes
-        self.method = method
-        self.state = TrainGenModelState(result_path, name, gen_examples_count)
-        self.reports = reports
 
-    def run(self, result_path: Path) -> TrainGenModelState:
+    def run(self, result_path: Path) -> GenModelState:
         self._set_path(result_path)
         self._fit_model()
         self._save_model()
@@ -98,43 +74,5 @@ class TrainGenModelInstruction(Instruction):
     def _save_model(self):
         self.state.model_path = self.method.save_model(self.state.result_path / 'trained_model/')
 
-    def _gen_data(self):
-        dataset = self.method.generate_sequences(self.state.gen_examples_count, 1,
-                                                 self.state.result_path / 'generated_sequences',
-                                                 SequenceType.AMINO_ACID, False)
-
-        print_log(f"{self.state.name}: generated {self.state.gen_examples_count} examples from the fitted model", True)
-        self.generated_dataset = dataset
-
-        AIRRExporter.export(dataset, self.state.result_path)
-
     def _evaluate_model(self):
         print("Evaluation is not implemented yet!")
-
-    def _run_reports(self):
-        report_path = PathBuilder.build(self.state.result_path / 'reports')
-        for report in self.reports:
-            report.result_path = report_path
-            if isinstance(report, DataReport):
-                tmp_report_synth = copy.deepcopy(report)
-                tmp_report_synth.name = tmp_report_synth.name + "_synthetic_data"
-                tmp_report_synth.dataset = self.generated_dataset
-                self.state.report_results['data_reports'].append(tmp_report_synth.generate_report())
-
-                tmp_report_org = copy.deepcopy(report)
-                tmp_report_org.name = tmp_report_org.name + '_original_data'
-                tmp_report_org.dataset = self.dataset
-                self.state.report_results['data_reports'].append(tmp_report_org.generate_report())
-
-            elif isinstance(report, GenModelReport):
-                report.model = self.method
-                report.dataset = self.dataset
-                self.state.report_results['ml_reports'].append(report.generate_report())
-
-        if len(self.reports) > 0:
-            gen_rep_count = len(self.state.report_results['ml_reports']) + int(
-                len(self.state.report_results['data_reports']) / 2)
-            print_log(f"{self.state.name}: generated {gen_rep_count} reports.", True)
-
-    def _set_path(self, result_path):
-        self.state.result_path = PathBuilder.build(result_path / self.state.name)
